+++ pycaw/api/mmdeviceapi/depend/__init__.py
@@ -1,5 +1,5 @@
 from ctypes import HRESULT, POINTER
-from ctypes.wintypes import DWORD, LPWSTR
+from ctypes.wintypes import DWORD
 
 from comtypes import COMMETHOD, GUID, IUnknown
 
@@ -7,25 +7,41 @@ from .structures import PROPERTYKEY, PROPVARIANT
 
 
 class IPropertyStore(IUnknown):
-    _iid_ = GUID('{886d8eeb-8cf2-4446-8d02-cdba1dbdcf99}')
+    _iid_ = GUID("{886d8eeb-8cf2-4446-8d02-cdba1dbdcf99}")
     _methods_ = (
         # HRESULT GetCount([out] DWORD *cProps);
-        COMMETHOD([], HRESULT, 'GetCount',
-                  (['out'], POINTER(DWORD), 'cProps')),
+        COMMETHOD([], HRESULT, "GetCount", (["out"], POINTER(DWORD), "cProps")),
         # HRESULT GetAt(
         # [in] DWORD iProp,
         # [out] PROPERTYKEY *pkey);
-        COMMETHOD([], HRESULT, 'GetAt',
-                  (['in'], DWORD, 'iProp'),
-                  (['out'], POINTER(PROPERTYKEY), 'pkey')),
+        COMMETHOD(
+            [],
+            HRESULT,
+            "GetAt",
+            (["in"], DWORD, "iProp"),
+            (["out"], POINTER(PROPERTYKEY), "pkey"),
+        ),
         # HRESULT GetValue(
         # [in] REFPROPERTYKEY key,
         # [out] PROPVARIANT *pv);
-        COMMETHOD([], HRESULT, 'GetValue',
-                  (['in'], POINTER(PROPERTYKEY), 'key'),
-                  (['out'], POINTER(PROPVARIANT), 'pv')),
-        # HRESULT SetValue([out] LPWSTR *ppstrId);
-        COMMETHOD([], HRESULT, 'SetValue',
-                  (['out'], POINTER(LPWSTR), 'ppstrId')),
+        COMMETHOD(
+            [],
+            HRESULT,
+            "GetValue",
+            (["in"], POINTER(PROPERTYKEY), "key"),
+            (["out"], POINTER(PROPVARIANT), "pv"),
+        ),
+        # HRESULT SetValue(
+        #   [in] REFPROPERTYKEY key,
+        #   [in] REFPROPVARIANT propvar
+        # );
+        COMMETHOD(
+            [],
+            HRESULT,
+            "SetValue",
+            (["in"], POINTER(PROPERTYKEY), "key"),
+            (["in"], POINTER(PROPVARIANT), "propvar"),
+        ),
         # HRESULT Commit();
-        COMMETHOD([], HRESULT, 'Commit'))
+        COMMETHOD([], HRESULT, "Commit"),
+    )