+++ pycaw/api/mmdeviceapi/depend/structures.py
@@ -1,6 +1,5 @@
 from ctypes import Structure, Union
-from ctypes.wintypes import (DWORD, LONG, LPWSTR, ULARGE_INTEGER, VARIANT_BOOL,
-                             WORD)
+from ctypes.wintypes import DWORD, LONG, LPWSTR, ULARGE_INTEGER, VARIANT_BOOL, WORD
 
 from comtypes import GUID
 from comtypes.automation import VARTYPE, VT_BOOL, VT_CLSID, VT_LPWSTR, VT_UI4
@@ -8,21 +7,21 @@ from comtypes.automation import VARTYPE, VT_BOOL, VT_CLSID, VT_LPWSTR, VT_UI4
 
 class PROPVARIANT_UNION(Union):
     _fields_ = [
-        ('lVal', LONG),
-        ('uhVal', ULARGE_INTEGER),
-        ('boolVal', VARIANT_BOOL),
-        ('pwszVal', LPWSTR),
-        ('puuid', GUID),
+        ("lVal", LONG),
+        ("uhVal", ULARGE_INTEGER),
+        ("boolVal", VARIANT_BOOL),
+        ("pwszVal", LPWSTR),
+        ("puuid", GUID),
     ]
 
 
 class PROPVARIANT(Structure):
     _fields_ = [
-        ('vt', VARTYPE),
-        ('reserved1', WORD),
-        ('reserved2', WORD),
-        ('reserved3', WORD),
-        ('union', PROPVARIANT_UNION),
+        ("vt", VARTYPE),
+        ("reserved1", WORD),
+        ("reserved2", WORD),
+        ("reserved3", WORD),
+        ("union", PROPVARIANT_UNION),
     ]
 
     def GetValue(self):
@@ -44,8 +43,8 @@ class PROPVARIANT(Structure):
 
 class PROPERTYKEY(Structure):
     _fields_ = [
-        ('fmtid', GUID),
-        ('pid', DWORD),
+        ("fmtid", GUID),
+        ("pid", DWORD),
     ]
 
     def __str__(self):