+++ pycaw/constants.py
@@ -2,11 +2,9 @@ from enum import Enum, IntEnum
 
 from comtypes import GUID
 
-IID_Empty = GUID(
-    '{00000000-0000-0000-0000-000000000000}')
+IID_Empty = GUID("{00000000-0000-0000-0000-000000000000}")
 
-CLSID_MMDeviceEnumerator = GUID(
-    '{BCDE0395-E52F-467C-8E3D-C4579291692E}')
+CLSID_MMDeviceEnumerator = GUID("{BCDE0395-E52F-467C-8E3D-C4579291692E}")
 
 
 class ERole(Enum):